/Z/tb-v/  table improvements -
Developed naming of table components.  In a table with headings across
the top and down the left, the top is line 0.sec, and the left is 1.Ti,
2.Ti, etc.  This is expressed as Z/tb-v/r/1 (2, 3, etc.) has a 0.sec
heading on the top.  Z/tb-v/r/s1 (2, 3, etc.) has no heading.
Z/tb-v/c/1 has heading on the left while Z/tb-v/c/s1 does not have a
heading on the left.